\name{charlson-package}
\alias{charlson-package}
\alias{charlson}
\docType{package}
\title{
Scoring Charlson using ICD9 codes
}
\description{
Given a list of patient id and comorbidities, this package will transpose the data to one row per patient and compute the Charlson score based on ICD9 codes
}
\details{
\tabular{ll}{
Package: \tab charlson\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-05-04\cr
License: \tab Unrestricted\cr
LazyLoad: \tab yes\cr
}
Overview:
Input data in two columns, first column is the patient identifier, the second column contains ICD9 codes.  The program will check for the ICD9 codes that count in the Charleson Index and compute for each patient.  
}
\author{
Vanessa Cox
Maintainer: Vanessa Cox
}
\references{
references
}
\keyword{ package }
\keyword{ charlson }
\keyword{ icd9 }
\keyword{ comorbidity }
